move examples into examples folder
