--- example/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "example"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ecrs = {path = "../" }
-itertools = "0.10.5"
-rand = "0.8.5"
-log4rs = "1.2.0"
-log = "0.4.17"
