+++ examples/ff.rs
@@ -2,7 +2,9 @@ use ecrs::ff::auxiliary::*;
 use ecrs::ff::probe::console_probe::ConsoleProbe;
 use ecrs::ff::*;
 
-pub fn firefly_example() {
+mod aco_tsp;
+
+fn main() {
   let mut alg = FireflyAlgorithm {
     config: FireflyAlgorithmCfg::default(),
     brightness_function: rastrigin,
