+++ examples/ga_bsc/wordmax.rs
